+++ docker/Transformation-Engine/src/main/java/nhsd/fhir/transformationenginepoc/service/converter/Transformer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nhsd.fhir.transformationenginepoc.service.transformers;
+package nhsd.fhir.transformationenginepoc.service.converter;
 
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.context.FhirVersionEnum;
