+++ docker/Transformation-Engine/src/test/java/nhsd/fhir/transformationenginepoc/service/transformers/BundleTransformerTest.java
@@ -1,6 +1,7 @@
 package nhsd.fhir.transformationenginepoc.service.transformers;
 
 import ca.uhn.fhir.context.FhirVersionEnum;
+import nhsd.fhir.transformationenginepoc.service.converter.BundleTransformer;
 import org.apache.commons.io.FileUtils;
 import org.custommonkey.xmlunit.XMLAssert;
 import org.custommonkey.xmlunit.XMLUnit;
